Add "Open in Colab" button for ease of use
Added a button at the top of the notebook to open it directly in Google Colab. This change is intended to make it easier for users to run the code without needing to set up a local environment. I initially added this for my own convenience and thought it might benefit others as well.